+++ cookidoo_api/helpers.py
@@ -8,15 +8,19 @@ from typing import cast
 
 from cookidoo_api.raw_types import (
     AdditionalItemJSON,
+    AuthResponseJSON,
     CustomCollectionJSON,
     IngredientJSON,
     ItemJSON,
     ManagedCollectionJSON,
     RecipeDetailsJSON,
     RecipeJSON,
+    SubscriptionJSON,
+    UserInfoJSON,
 )
 from cookidoo_api.types import (
     CookidooAdditionalItem,
+    CookidooAuthResponse,
     CookidooCategory,
     CookidooChapter,
     CookidooChapterRecipe,
@@ -27,11 +31,53 @@ from cookidoo_api.types import (
     CookidooRecipeCollection,
     CookidooShoppingRecipe,
     CookidooShoppingRecipeDetails,
+    CookidooSubscription,
+    CookidooUserInfo,
 )
 
 _LOGGER = logging.getLogger(__name__)
 
 localization_file_path = os.path.join(os.path.dirname(__file__), "localization.json")
+
+
+def cookidoo_auth_data_from_json(
+    auth_data: AuthResponseJSON,
+) -> CookidooAuthResponse:
+    """Convert a auth data received from the API to a cookidoo auth data."""
+    return CookidooAuthResponse(
+        username=auth_data["user_name"],
+        access_token=auth_data["access_token"],
+        refresh_token=auth_data["refresh_token"],
+        token_type=auth_data["token_type"],
+        expires_in=auth_data["expires_in"],
+    )
+
+
+def cookidoo_user_info_from_json(
+    user_info: UserInfoJSON,
+) -> CookidooUserInfo:
+    """Convert a user info received from the API to a cookidoo user info."""
+    return CookidooUserInfo(
+        username=user_info["username"],
+        description=user_info["description"],
+        picture=user_info["picture"],
+    )
+
+
+def cookidoo_subscription_from_json(
+    subscription: SubscriptionJSON,
+) -> CookidooSubscription:
+    """Convert a subscription received from the API to a cookidoo subscription."""
+    return CookidooSubscription(
+        active=subscription["active"],
+        expires=subscription["expires"],
+        start_date=subscription["startDate"],
+        status=subscription["status"],
+        subscription_level=subscription["subscriptionLevel"],
+        subscription_source=subscription["subscriptionSource"],
+        type=subscription["type"],
+        extended_type=subscription["extendedType"],
+    )
 
 
 def cookidoo_collection_from_json(
@@ -67,7 +113,7 @@ def cookidoo_recipe_from_json(
         id=recipe["id"],
         name=recipe["title"],
         ingredients=[
-            cookidoo_ingredient_item_from_json(ingredient)
+            cookidoo_ingredient_from_json(ingredient)
             for ingredient in recipe["recipeIngredientGroups"]
         ],
     )
@@ -120,11 +166,11 @@ def cookidoo_recipe_details_from_json(
 
 
 def cookidoo_ingredient_from_json(
-    ingredient: IngredientJSON,
+    ingredient: IngredientJSON | ItemJSON,
 ) -> CookidooIngredient:
     """Convert an ingredient received from the API to a cookidoo ingredient."""
     return CookidooIngredient(
-        id=ingredient["localId"],
+        id=ingredient["localId"] if "localId" in ingredient else ingredient["id"],  # type: ignore[typeddict-item]
         name=ingredient["ingredientNotation"],
         description=f"{ingredient['quantity']['value']} {ingredient['unitNotation']}"
         if ingredient["unitNotation"] and ingredient["quantity"]
@@ -166,14 +212,14 @@ def __get_localization_options(
     language: str | None = None,
 ) -> list[CookidooLocalizationConfig]:
     with open(localization_file_path, encoding="utf-8") as file:
-        options = cast(list[CookidooLocalizationConfig], json.loads(file.read()))
-        return list(
-            filter(
-                lambda option: (not country or option["country_code"] == country)
-                and (not language or option["language"] == language),
-                options,
-            )
+        options_ = cast(list[dict[str, str]], json.loads(file.read()))
+        options = (CookidooLocalizationConfig(**x) for x in options_)
+        filtered_options = filter(
+            lambda option: (not country or option.country_code == country)
+            and (not language or option.language == language),
+            options,
         )
+        return list(cast(list[CookidooLocalizationConfig], filtered_options))
 
 
 async def get_localization_options(
@@ -188,9 +234,9 @@ async def get_localization_options(
 
 async def get_country_options() -> list[str]:
     """Get a list of possible country options."""
-    return list({option["country_code"] for option in await get_localization_options()})
+    return list({option.country_code for option in await get_localization_options()})
 
 
 async def get_language_options() -> list[str]:
     """Get a list of possible language options."""
-    return list({option["language"] for option in await get_localization_options()})
+    return list({option.language for option in await get_localization_options()})